requirements.txt
annotated-types==0.7.0
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
gitdb==4.0.11
GitPython==3.1.41
idna==3.7
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.7.3
pydantic-settings==2.3.1
pydantic_core==2.18.4
pymongo==4.7.3
python-dotenv==1.0.1
requests==2.32.3
setuptools==69.0.3
smmap==5.0.1
typing_extensions==4.12.1
urllib3==2.2.1
wheel==0.43.0
